add function for highcharts

+++ db/migrate/20110826020211_create_rates.rb
@@ -3,7 +3,7 @@ class CreateRates < ActiveRecord::Migration
     create_table :rates do |t|
       t.integer :bank_id
       t.integer :category_rate_type_id
-      t.decimal :percent_rate
+      t.float :percent_rate
 
       t.timestamps
     end
